+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/dialogs/AltColorUploadProgressModel.java
@@ -68,7 +68,7 @@ public class AltColorUploadProgressModel extends ProgressModel<File> {
       }
       else {
         Platform.runLater(() -> {
-          EventManager.getInstance().notifyJobFinished(ALTCOLOR_INSTALL);
+          EventManager.getInstance().notifyJobFinished(ALTCOLOR_INSTALL, gameId);
         });
       }
       progressResultModel.addProcessed();