+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/dialogs/DirectB2SUploadProgressModel.java
@@ -66,7 +66,7 @@ public class DirectB2SUploadProgressModel extends ProgressModel<File> {
       }
       else {
         Platform.runLater(() -> {
-          EventManager.getInstance().notifyJobFinished(DIRECTB2S_INSTALL);
+          EventManager.getInstance().notifyJobFinished(DIRECTB2S_INSTALL, gameId);
         });
       }
       progressResultModel.addProcessed();